+++ Haden.NxtSharp/DataPoints.cs
@@ -1,7 +1,0 @@
-﻿namespace Haden.NxtSharp
-{
-    public class DataPoints
-    {
-        public enum EventDataPoint { Compute, Control, Seek, Sense }
-    }
-}