+++ Haden.NxtSharp/Extensions.cs
@@ -1,8 +1,0 @@
-﻿
-namespace Haden.NxtSharp
-{
-    public static class Extensions
-    {
-
-    }
-}