+++ Haden.NxtSharp/Sensors/SensorEvent.cs
@@ -1,7 +1,0 @@
-namespace Haden.NxtSharp.Sensors
-{
-    /// <summary>
-    /// The SensorEvent delegate.
-    /// </summary>
-    public delegate void SensorEvent(NxtSensor sensor);
-}